Solve Firefox :has(:focus-visible) bug by switching to JavaScript
Using the :has(:focus-visible) CSS declaration to keep the parent
container visible while switching the focus from the parent to the children caused issues on Firefox, as it removed the hover and focus effects altogether. The bug was solved by using the :focus-within declaration on the parent while preventing the default click behavior of the children using JavaScript. This ensures that clicking the children will not cause the parent container to remain stuck in the "focus" state, allowing for it to return to its initial transform value.